Part 1
Separating codbase to shared library
