--- Client/CustomResumeBlazor.csproj
@@ -1,24 +1,0 @@
-﻿<Project Sdk="Microsoft.NET.Sdk.BlazorWebAssembly">
-
-	<PropertyGroup>
-		<TargetFramework>net8.0</TargetFramework>
-		<Nullable>enable</Nullable>
-		<ImplicitUsings>enable</ImplicitUsings>
-		<WasmEnableSIMD>false</WasmEnableSIMD>
-		<WasmEnableExceptionHandling>false</WasmEnableExceptionHandling>
-	</PropertyGroup>
-
-	<ItemGroup>
-		<PackageReference Include="Blazored.LocalStorage" Version="4.4.0" />
-		<PackageReference Include="Microsoft.AspNetCore.Components.WebAssembly" Version="8.0.1" />
-		<PackageReference Include="Microsoft.AspNetCore.Components.WebAssembly.DevServer" Version="8.0.1" PrivateAssets="all" />
-		<PackageReference Include="MudBlazor" Version="6.14.0" />
-	</ItemGroup>
-
-	<ItemGroup>
-	  <Content Update="wwwroot\database\websiteData.json">
-	    <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
-	  </Content>
-	</ItemGroup>
-	
-</Project>
