--- Client/Domain/Mappers/RouteMapper.cs
@@ -1,7 +1,0 @@
-﻿namespace CustomResumeBlazor.Domain.Mappers;
-
-public static class RouteMapper
-{
-    public static string GetAllCardsRoute(string endpoint) => $"defaultviewer/{endpoint}";
-}
-
