+++ Custom-Resume.sln
@@ -1,0 +1,31 @@
+﻿
+Microsoft Visual Studio Solution File, Format Version 12.00
+# Visual Studio Version 17
+VisualStudioVersion = 17.0.31903.59
+MinimumVisualStudioVersion = 10.0.40219.1
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "CustomResume.Blazor", "CustomResume.Blazor\CustomResume.Blazor.csproj", "{4A912E35-2585-41D3-A0D9-DDB6727426B9}"
+EndProject
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "CustomResume.Library", "CustomResume.Library\CustomResume.Library.csproj", "{35E0A33A-FB6D-400C-AE6A-2F76C1C39646}"
+EndProject
+Global
+	GlobalSection(SolutionConfigurationPlatforms) = preSolution
+		Debug|Any CPU = Debug|Any CPU
+		Release|Any CPU = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(ProjectConfigurationPlatforms) = postSolution
+		{4A912E35-2585-41D3-A0D9-DDB6727426B9}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{4A912E35-2585-41D3-A0D9-DDB6727426B9}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{4A912E35-2585-41D3-A0D9-DDB6727426B9}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{4A912E35-2585-41D3-A0D9-DDB6727426B9}.Release|Any CPU.Build.0 = Release|Any CPU
+		{35E0A33A-FB6D-400C-AE6A-2F76C1C39646}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{35E0A33A-FB6D-400C-AE6A-2F76C1C39646}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{35E0A33A-FB6D-400C-AE6A-2F76C1C39646}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{35E0A33A-FB6D-400C-AE6A-2F76C1C39646}.Release|Any CPU.Build.0 = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(SolutionProperties) = preSolution
+		HideSolutionNode = FALSE
+	EndGlobalSection
+	GlobalSection(ExtensibilityGlobals) = postSolution
+		SolutionGuid = {B98459E4-DA48-456F-AF12-D1208270E92C}
+	EndGlobalSection
+EndGlobal
