+++ CustomResume.Blazor/.editorconfig
