+++ CustomResume.Blazor/App.razor
@@ -1,3 +1,5 @@
+@using CustomResume.Blazor.Shared
+
 <Router AppAssembly="@typeof(App).Assembly">
     <Found Context="routeData">
         <RouteView RouteData="@routeData" DefaultLayout="@typeof(MainLayout)" />
