+++ CustomResume.Blazor/Pages/AllCardsPage.razor
@@ -1,0 +1,4 @@
+﻿@page "/card/{ClientRouteName}"
+@using CustomResume.Library.Application.Components
+
+<DisplayAllCardsPage ClientRouteName="@ClientRouteName" />
