+++ CustomResume.Blazor/Pages/AllCardsPage.razor.cs
@@ -1,0 +1,9 @@
+using Microsoft.AspNetCore.Components;
+
+namespace CustomResume.Blazor.Pages;
+
+public partial class AllCardsPage
+{
+    [Parameter]
+    public string ClientRouteName { get; set; } = default!;
+}
