+++ CustomResume.Blazor/Pages/Index.razor
@@ -1,0 +1,5 @@
+﻿@page "/"
+@using CustomResume.Library.Application
+@using CustomResume.Library.Application.Components
+
+<LandingPage />
