+++ CustomResume.Blazor/Shared/NavMenu.razor
@@ -27,3 +27,4 @@
     }
 
 </MudNavMenu>
+
