+++ CustomResume.Blazor/_Imports.razor
@@ -7,6 +7,5 @@
 @using Microsoft.AspNetCore.Components.WebAssembly.Http
 @using Microsoft.JSInterop
 @using MudBlazor
-@using CustomResumeBlazor
-@using CustomResumeBlazor.Shared
+@using CustomResume.Blazor
