+++ CustomResume.Blazor/wwwroot/css/open-iconic/FONT-LICENSE
