+++ CustomResume.Blazor/wwwroot/css/open-iconic/README.md
