+++ CustomResume.Blazor/wwwroot/css/open-iconic/font/fonts/open-iconic.ttf
