+++ CustomResume.Library/Application/Components/DisplayAllCardsPage.razor
@@ -1,10 +1,11 @@
-﻿@page "/defaultviewer/{ClientRouteName}"
-@using CustomResumeBlazor.Domain
+﻿@using CustomResumeBlazor.Domain
 @using CustomResumeBlazor.Domain.Mappers
+@using CustomResumeBlazor.Infrastructure
+
 
 <PageTitle>@ClientRouteName.FirstCharToUpper()</PageTitle>
 
-    @if (_hasLoaded is false)
+@if (_hasLoaded is false)
 {
     <MudAlert Severity="Severity.Info" Variant="Variant.Filled" Class="m-4">
         <MudAlert>Page is loading...</MudAlert>
@@ -19,13 +20,13 @@ else if (_currentPage?.PageFormat is CardType.Experience or CardType.Project or 
                 <MudCard Square="true" Style="width:400px">
                     <MudCardContent>
                         <MudText Typo="Typo.h5">@project.Title</MudText>
-                            <MudText Typo="Typo.body2">@project.Body</MudText>
-                        </MudCardContent>
-                        <MudCardActions>
-                            <MudButton Variant="Variant.Text" Color="Color.Primary" Href="@project.LearnMoreUrl" Target="_blank">Learn More</MudButton>
-                        </MudCardActions>
-                    </MudCard>
-                </MudItem>
+                        <MudText Typo="Typo.body2">@project.Body</MudText>
+                    </MudCardContent>
+                    <MudCardActions>
+                        <MudButton Variant="Variant.Text" Color="Color.Primary" Href="@project.LearnMoreUrl" Target="_blank">Learn More</MudButton>
+                    </MudCardActions>
+                </MudCard>
+            </MudItem>
         }
     </MudItem>
 }
@@ -43,11 +44,12 @@ else if (_currentPage?.PageFormat is CardType.Skill)
                         <MudIcon Icon="@IconMapper.GetIcon(skill.Icon)" Size="Size.Large" />
                         <MudCardContent>
                             <MudText Typo="Typo.h5">@skill.Title</MudText>
-                                <MudText Typo="Typo.body2">@skill.Body</MudText>
-                            </MudCardContent>
-                        </MudCard>
-                    </MudItem>
+                            <MudText Typo="Typo.body2">@skill.Body</MudText>
+                        </MudCardContent>
+                    </MudCard>
+                </MudItem>
             }
         </MudItem>
     }
 }
+
