+++ CustomResume.Library/CustomResume.Library.csproj
@@ -1,0 +1,17 @@
+﻿<Project Sdk="Microsoft.NET.Sdk.Razor">
+
+	<PropertyGroup>
+		<TargetFramework>net8.0</TargetFramework>
+		<Nullable>enable</Nullable>
+		<ImplicitUsings>enable</ImplicitUsings>
+		<WasmEnableSIMD>false</WasmEnableSIMD>
+		<WasmEnableExceptionHandling>false</WasmEnableExceptionHandling>
+	</PropertyGroup>
+
+	<ItemGroup>
+		<PackageReference Include="Blazored.LocalStorage" Version="4.5.0" />
+		<PackageReference Include="Microsoft.AspNetCore.Components.Web" Version="8.0.4" />
+		<PackageReference Include="MudBlazor" Version="6.19.1" />
+	</ItemGroup>
+
+</Project>
