+++ CustomResume.Library/Domain/Mappers/IconMapper.cs
