+++ CustomResume.Library/Domain/Mappers/RouteMapper.cs
@@ -1,0 +1,7 @@
+﻿namespace CustomResumeBlazor.Domain.Mappers;
+
+public static class RouteMapper
+{
+    public static string GetAllCardsRoute(string endpoint) => $"card/{endpoint}";
+}
+
