+++ CustomResume.Library/Domain/MiExtensions.cs
